requirements.txt
psycopg2-binary
python-dotenv
google-generativeai
python-telegram-bot==20.3
